+++ plugins/json-extractor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flatfile/plugin-json-extractor
 
+## 0.4.3
+
+### Patch Changes
+
+- @flatfile/util-extractor@0.2.2
+
 ## 0.4.2
 
 ### Patch Changes