+++ plugins/zip-extractor/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @flatfile/plugin-zip-extractor
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [490e8f1]
+  - @flatfile/util-file-buffer@0.0.3
+
+## 0.3.2
+
+### Patch Changes
+
+- 010a31c: Exclude \_\_MACOSX
+
 ## 0.3.1
 
 ### Patch Changes